feat: refactor the environments module
