+++ src/pdm/cli/actions.py
@@ -34,13 +34,13 @@ from pdm.cli.utils import (
     save_version_specifiers,
     set_env_in_reg,
 )
+from pdm.environments import BareEnvironment, PythonLocalEnvironment
 from pdm.exceptions import NoPythonVersion, PdmUsageError, ProjectError
 from pdm.formats import FORMATS
 from pdm.formats.base import array_of_inline_tables, make_array, make_inline_table
 from pdm.models.backends import DEFAULT_BACKEND, BuildBackend
 from pdm.models.caches import JSONFileCache
 from pdm.models.candidates import Candidate
-from pdm.models.environment import BareEnvironment
 from pdm.models.python import PythonInfo
 from pdm.models.requirements import Requirement, parse_requirement, strip_extras
 from pdm.models.specifiers import get_specifier
@@ -643,7 +643,11 @@ def do_use(
         f"Using Python interpreter: [success]{str(selected_python.path)}[/] ({selected_python.identifier})"
     )
     project.python = selected_python
-    if old_python and old_python.executable != selected_python.executable and not project.environment.is_global:
+    if (
+        old_python
+        and old_python.executable != selected_python.executable
+        and isinstance(project.environment, PythonLocalEnvironment)
+    ):
         project.core.ui.echo("Updating executable scripts...", style="primary")
         project.environment.update_shebangs(selected_python.executable.as_posix())
     hooks.try_emit("post_use", python=selected_python)
