+++ src/pdm/cli/commands/run.py
@@ -152,19 +152,7 @@ class TaskRunner:
         pythonpath = [get_pep582_path(project)] + [p for p in pythonpath if "pdm/pep582" not in p.replace("\\", "/")]
         project_env = project.environment
         this_path = project_env.get_paths()["scripts"]
-        python_root = os.path.dirname(project.python.executable)
-        new_path = os.pathsep.join([this_path, process_env.get("PATH", ""), python_root])
-        process_env.update(
-            {
-                "PYTHONPATH": os.pathsep.join(pythonpath),
-                "PATH": new_path,
-                "PDM_PROJECT_ROOT": str(project.root),
-            }
-        )
-        if project_env.packages_path:
-            process_env.update({"PEP582_PACKAGES": str(project_env.packages_path)})
-        if project_env.venv_path:
-            process_env.update({"VIRTUAL_ENV": str(project_env.venv_path)})
+        process_env.update(project_env.process_env)
         if env:
             process_env.update(env)
         if shell:
@@ -183,7 +171,7 @@ class TaskRunner:
             real_command = os.path.realpath(expanded_command)
             expanded_args = [os.path.expandvars(arg) for arg in [expanded_command, *args]]
             if (
-                not project_env.is_global
+                project_env.is_local
                 and not site_packages
                 and (
                     os.path.basename(real_command).startswith("python")
