+++ src/pdm/models/python.py
@@ -8,6 +8,7 @@ from findpython import PythonVersion
 from packaging.version import InvalidVersion, Version
 
 from pdm.compat import cached_property
+from pdm.utils import get_venv_like_prefix
 
 
 class PythonInfo:
@@ -78,3 +79,7 @@ class PythonInfo:
             return f"{self.major}.{self.minor}"
         except InvalidVersion:
             return "unknown"
+
+    @property
+    def is_venv(self) -> bool:
+        return get_venv_like_prefix(self.executable) is not None
