+++ src/pdm/pytest.py
@@ -53,11 +53,11 @@ from pdm.cli.actions import do_init
 from pdm.cli.hooks import HookManager
 from pdm.compat import Protocol
 from pdm.core import Core
+from pdm.environments import BaseEnvironment, PrefixEnvironment
 from pdm.exceptions import CandidateInfoNotFound
 from pdm.installers.installers import install_wheel
 from pdm.models.backends import get_backend
 from pdm.models.candidates import Candidate
-from pdm.models.environment import Environment, PrefixEnvironment
 from pdm.models.repositories import BaseRepository
 from pdm.models.requirements import (
     Requirement,
@@ -319,7 +319,7 @@ def build_env(build_env_wheels: Iterable[Path], tmp_path_factory: pytest.TempPat
     """
     d = tmp_path_factory.mktemp("pdm-test-env")
     p = Core().create_project(d)
-    env = PrefixEnvironment(p, str(d))
+    env = PrefixEnvironment(p, prefix=str(d))
     for wheel in build_env_wheels:
         install_wheel(str(wheel), env)
     return d
@@ -437,7 +437,7 @@ def working_set(mocker: MockerFixture, repository: TestRepository) -> MockWorkin
         a mock working set
     """
     rv = MockWorkingSet()
-    mocker.patch.object(Environment, "get_working_set", return_value=rv)
+    mocker.patch.object(BaseEnvironment, "get_working_set", return_value=rv)
 
     def install(candidate: Candidate) -> None:
         key = normalize_name(candidate.name or "")
