+++ tests/test_project.py
@@ -7,6 +7,7 @@ import pytest
 from packaging.version import parse
 
 from pdm.cli.commands.venv.utils import get_venv_python
+from pdm.environments import PythonEnvironment
 from pdm.exceptions import PdmException
 from pdm.utils import cd
 
@@ -94,7 +95,8 @@ def test_project_sources_env_var_expansion(project, monkeypatch):
 
 def test_global_project(tmp_path, core):
     project = core.create_project(tmp_path, True)
-    assert project.environment.is_global
+    assert project.is_global
+    assert isinstance(project.environment, PythonEnvironment)
 
 
 def test_auto_global_project(tmp_path, core):
@@ -137,7 +139,7 @@ def test_project_auto_detect_venv(project):
     project._python = None
     project._saved_python = (project.root / "test_venv" / scripts / f"python{suffix}").as_posix()
 
-    assert project.environment.is_global
+    assert not project.environment.is_local
 
 
 @pytest.mark.path
